Update README.md
Newsgroup Text Clustering
This repository contains code for clustering newsgroup text documents using a variety of clustering algorithms. The code is written in Python and uses the NLTK and Scikit-learn libraries.

The code will first read in a directory of newsgroup text files. It will then create a TF-IDF vector for each document. Finally, it will cluster the documents using a variety of clustering algorithms. The results of the clustering will be printed to the console.

Further Work
The code could be further improved by using a more sophisticated clustering algorithm, such as a hierarchical clustering algorithm. Additionally, the code could be extended to cluster documents in other languages.